+++ CHANGELOG.md
@@ -18,6 +18,7 @@ the section headers (Added/Changed/...) and incrementing the package version.
 ### Changed
 
 ### Fixed
+- ([#161](https://github.com/microsoft/hi-ml/pull/161)) Empty string as target folder for a dataset creates an invalid mounting path for the dataset in AzureML (fixes #160)
 
 ### Removed